restructured codebase again
