+++ src/components/graph.js
@@ -78,93 +78,9 @@ const Graph = ({ data, scheduler = false }) => {
 
 export default Graph;
 
-
 // util functions for graph and gpa
-
-// Helper to sanitize description
 export const sanitizeDescription = (data) => {
 	if (data.description && data.description.startsWith("<a href=")) {
 		data.description = "No Description Available";
 	}
 };
-
-// Helper to collect all professors
-export const collectAllProfessors = (instructors) => {
-	const allProfs = [];
-	for (const semester in instructors) {
-		for (const instructor of instructors[semester]) {
-			if (!allProfs.includes(instructor)) {
-				allProfs.push(instructor);
-			}
-		}
-	}
-	return allProfs;
-};
-
-// Helper to calculate GPA and grade distributions
-export const calculateGradesAndGPA = (profs, gpaData, colors) => {
-	const grades = [];
-	const gpa = {};
-	let colorIndex = 0;
-
-	for (const instructor of profs) {
-		let avgGPA = 0;
-		let avgGradeDist = Array(13).fill(0);
-		const color = colors[colorIndex++ % colors.length];
-
-		if (!gpaData[instructor]) {
-			gpa[instructor] = [0, "#ffffff"];
-			grades.push({
-				label: instructor,
-				data: avgGradeDist,
-				backgroundColor: "#ffffff",
-			});
-			continue;
-		}
-
-		let semesterCount = 0;
-		for (const sem in gpaData[instructor]) {
-			avgGPA += gpaData[instructor][sem][13];
-			avgGradeDist = avgGradeDist.map(
-				(val, i) => val + gpaData[instructor][sem][i]
-			);
-			semesterCount++;
-		}
-
-		avgGradeDist = avgGradeDist.map((val) =>
-			Math.round((val / semesterCount) * 100) / 100
-		);
-
-		gpa[instructor] = [
-			Math.round((avgGPA / semesterCount) * 100) / 100,
-			color,
-		];
-		grades.push({
-			label: instructor,
-			data: avgGradeDist,
-			backgroundColor: color,
-		});
-	}
-	return { grades, gpa };
-};
-
-
-import { graphColors } from '@/lib/utils';
-export const averageAllData = (grades) => {
-
-	const avg = Array(13).fill(0);
-	for (const grade of grades) {
-		grade.data.forEach((val, i) => {
-			avg[i] += val;
-		});
-	}
-
-	const avgData = [{
-		label: "Average",
-		data: avg.map((val) => Math.round(val / grades.length * 100) / 100),
-		backgroundColor: graphColors[2],
-	}];
-
-	return avgData;
-
-}
